Refactored to clean up common code across lin,mac,win
