+++ contributed/cli/manifest.json
@@ -1,28 +1,8 @@
 {
-	"platforms": {
-		"x-cli-win": {
-			"modules": {
-				"*": [
-					"./x-cli-win/main",
-					"./x-cli-win/machine"
-				]
-			}
-		},
-		"x-cli-mac": {
-			"modules": {
-				"*": [
-					"./x-cli-mac/main",
-					"./x-cli-mac/machine"
-				]
-			}
-		},
-		"x-cli-lin": {
-			"modules": {
-				"*": [
-					"./x-cli-lin/main",
-					"./x-cli-lin/machine"
-				]
-			}
-		}
+	"modules": {
+		"*": [
+			"./$(PLATFORM)/main",
+			"./cli"
+		]
 	}
 }
