--- contributed/cli/x-cli-win/machine.h
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2020 Chris Midgley
- *
- *   This file is part of the Moddable SDK Tools.
- * 
- *   The Moddable SDK Tools is free software: you can redistribute it and/or modify
- *   it under the terms of the GNU General Public License as published by
- *   the Free Software Foundation, either version 3 of the License, or
- *   (at your option) any later version.
- * 
- *   The Moddable SDK Tools is distributed in the hope that it will be useful,
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *   GNU General Public License for more details.
- * 
- *   You should have received a copy of the GNU General Public License
- *   along with the Moddable SDK Tools.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
- 
-int startMachine(char *archivePath);
-void endMachine();
-void instrumentMachine();
-
